Move migrating RefFile to RefJSONFile into dedicated function
